requirements.txt
aniso8601==7.0.0
astroid==2.3.1
autopep8==1.4.4
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.2.0
django-graphql-jwt==0.2.2
django-heroku==0.3.1
djangorestframework==3.10.3
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
#pkg-resources==0.0.0
promise==2.2.1
psycopg2==2.8.3
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.2
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.4
wrapt==1.11.2